Rename response builder and parser
