+++ coordinator/src/services/communication/tracker/spine-response-parser.ts
